+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - `Ui::reload_stylesheet` added.
 - Loader system has been refactored
 - Margins added to stylesheet system. Margins automatically handled for all widgets.
+- Added `widget::input::State::set_value`
+- Removed scrollbars from stylesheet in favor of the new `Dummy` widget.
 
  ### v0.3.0
 - Added `len()` to `Widget`.